requirements.txt
colorama==0.4.6
mysql-connector-python==9.0.0
validate-docbr==1.10.0